third_party/WebKit/LayoutTests/fast/dom/shadow/offsetWidth-host-style-change-expected.txt

CONSOLE WARNING: line 12: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
offsetWidth of a fixed width element should cause a style recalc if host styles are invalid

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS div.offsetWidth is 100
PASS div.offsetWidth is 0
